splitting into components
