+++ css/common.css
@@ -19,14 +19,10 @@
 
 /* Every page */
 
-h1 {
-    text-align: center;
-}
+/* start by hiding everything */
+#welcome-page{display: none;}
 
-h2 {
-    color: var(--citygreen);
-    text-align: center;
-}
+#prompt-div{display: none;}
 
 body {
     margin-top: 20px;
