+++ index.html
@@ -7,63 +7,83 @@
 
 <!-- Bootstrap -->
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" crossorigin="anonymous">
-<!-- CSS -->
+<!-- common CSS -->
 <link rel="stylesheet" href="css/common.css">
-<link rel="stylesheet" href="css/00_welcome.css">
+<!-- Component CSS -->
+<link rel="stylesheet" href="js/components/welcome/welcome.css">
+<link rel="stylesheet" href="js/components/header/header.css">
+<link rel="stylesheet" href="js/components/prompt/prompt.css">
 
 <!-- Bootstrap JS and its dependencies (jQuery & Popper.js) -->
 <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/popper.js@1.9.4/dist/umd/popper.min.js" crossorigin="anonymous"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" crossorigin="anonymous"></script>
 
+<!-- My javascript -->
+<script src="js/init.js" type="module"></script>
+
 </head>
 <body>
 
 <h1>FY2026 Budget Form</h1>
+<h2 id="subtitle"></h2>
 
 <br>
 <br>
 <br>
 
 <div id="welcome-page">
-    
-    <a href="">
+    <a>
         <button class="btn step" id="step-upload">
             1. Upload your Excel data
         </button>
     </a>
     
-    <a href="pages/02_new_initiatives.html">
+    <a>
         <button class="btn step" id="step-initiatives">
             2. Name any new initiatives
         </button>
     </a>
 
-    <a href="">
+    <a>
         <button class="btn step" id="step-revenue">
             3. Confirm revenue
         </button>
     </a>
 
-    <a href="">
+    <a>
         <button class="btn step" id="step-personnel">
             4. Personnel costs
         </button>
     </a>
 
-    <a href="">
+    <a>
         <button class="btn step" id="step-nonpersonnel">
             5. Non-personnel costs
         </button>
     </a>
 
-    <a href="">
+    <a>
         <button class="btn step" id="step-finish">
             6. Confirm and finish
         </button>
     </a>
-    
 </div>
+
+<div id="prompt-div">
+    <h3 id="prompt">Do you have any new initiatives for FY26?</h3>
+    <br>
+        <button class="btn btn-yes" id="option1" data-toggle="modal" data-target="#new-init-modal">
+            Yes
+        </button>
+    <a href="../pages/03_revenue.html">
+        <button class="btn btn-no" id="option2">
+            No
+        </button>
+    </a>
+</div>
+
+
 
 </body>
 </html>
