+++ js/components/header/header.css
@@ -1,0 +1,9 @@
+
+h1 {
+    text-align: center;
+}
+
+h2 {
+    color: var(--citygreen);
+    text-align: center;
+}
