+++ js/components/header/header.js
@@ -1,0 +1,3 @@
+export function updateSubtitle(subtitle){
+    document.getElementById("subtitle").textContent = subtitle;
+}
