+++ js/components/prompt/prompt.js
@@ -1,0 +1,3 @@
+export function showPrompt(){
+    document.getElementById("prompt-div").style.display = "block";
+}
