+++ js/components/welcome/welcome.js
@@ -1,0 +1,7 @@
+// Hide and unhide welcome buttons
+export function unhideWelcomeButtons(){
+    document.getElementById("welcome-page").style.display = "flex";
+}
+export function hideWelcomeButtons(){
+    document.getElementById("welcome-page").style.display = "none";
+}
