+++ js/init.js
@@ -8,3 +8,14 @@ let baseline_revenue = 0;
 let supp_revenue = 0;
 let supp_total = personnel_supp - supp_revenue;
 let baseline_total = personnel_baseline - baseline_revenue;
+
+// import functions
+import { initializeWelcomePage } from './pages/00_welcome/main.js';
+import { loadNewInitiatives } from './pages/02_new_initiatives/main.js'
+
+document.addEventListener('DOMContentLoaded', function () {
+
+    initializeWelcomePage();
+    document.getElementById('step-initiatives').addEventListener('click', loadNewInitiatives)
+
+});
