+++ pages/02_new_initiatives.html
@@ -25,13 +25,16 @@
 <body>
 
 <h1>FY2026 Budget Form</h1>
-<h2>New Initiatives</h2>
+<h2 id="section">New Initiatives</h2>
 
 <br>
 <br>
 <br>
 
-<div id="welcome-page">
+
+
+
+<div id="new-inits">
     
     <!-- initial questions to hide once selected -->
     <div id="initial-questions">
